Fixed a performance issue with PathNode (#33)
PathNode.PathContains was generating the entire path to the root before iterating to test for equality, lowering the performance in cases when the given state can be found near the beginning of the collection. I fixed this by using GetPathToRoot instead of GetPath (the latter contains a Reverse call on the path IEnumerable).